aiohttp-giphy-navigator/requirements.txt
dependency-injector
aiohttp
pyyaml
pytest-aiohttp
pytest-cov
pytest-asyncio
